change: use pygame.system.get_pref_path()
To define a writeable path for things like configuration, etc.

It now replaces the custom makeshift solution we were using.

This solution now stores files holding data like configuration,
recent files, bookmarked folders in the file browser, etc., in
a new location.

Proper mechanisms were also employed to guarantee data in the
old locations is copied into the new locations in case the
files don't exist in the new locations. This way configuration
and other important data that exist prior to this change are not
lost.