add: birds eye view
New feature that allows users to visualize a representation of the
entire graph that fits the screen. In other words, it allows users
to see the whole graph at once.

The feature can also be used for navigation. Keeping the mouse pressed
while hovering the representation will cause the screen to scroll to
the respective spot in the actual graph. That is, that spot will appear
centered on the screen.

Because this new feature supersedes the one that allow users to jump to
different corners of the graph, the superseded feature was removed.

The general controls page was also updated to reflect the changes.