fix: suppress message when importing pygame-ce
Pygame-ce is already properly credited in the app by displaying its
logo on the splash screen with a link to its website. So there's no
harm in suppressing this message.

Moreover, we only suppress it to leave the stdout clean for when we
are launching the app repeatedly during debugging.