change: improve file selection on file browser
This was done by creating the "expecting_files_only" flag given when
using the file browser to select paths. By default it is set to False.

When True, operations involving folders do not change the selection
entry. This allows the user to freely select a folder either by
keyboard or mouse while navigating without having the selection
change the text on the selection entry.

Other changes were also made that are not directly related to this
flag we introduced but which also contribute to our goal of leaving
the text on the selection entry untouched.