change: improve form to export images
Name of loaded file is now used as the default name of the image
file to be exported when the user opens the form for the first time
after loading a file.

Also, when opening the file browser for selecting a new location
or changing the name, the selection entry is automatically filled
with the current file name using the current extension, instead of
defaulting to .html.