change: improve menu
Defining and controlling widgets in menus is a bit easier now, cause
reading and setting values is done with callables and radiobuttons
can be defined as a single menu item which is used to generate
multiple items in a preprocessing step.

Radiobutton surfaces are now generated with SVG, providing more
control over the final result.

When a menu has both a command with key text and a submenu, we only
use extra space for the element with largest width, instead of taking
both into account as before.

This makes the menu more compact and without harming readability.
Taking both into account was a waste of space as well, since submenus
don't use key text (as far as I know).