add: socket detection feature
This feature significantly changes how users interact with sockets
when creating/replacing connections, both when interacting with an
origin socket (the socket from where we drag a new connection) and
when interacting with a target socket (the socket where we drop a
new connection).

Before this change, whenever we wanted to create a new connection
between sockets, we had to:

- click a socket and drag a new connection from it
- drop the connection exactly over the other socket

Now doing both things is easier, because the mouse doesn't need to
be exactly over the sockets for any of the actions to take place.

In the first place, the user doesn't need to hover the mouse exactly
over the origin socket before clicking it to drag a new connection
from it: as long as the mouse is close enough to the socket, even
if the user clicks and drags an empty spot on the canvas, a new
connection starts being defined right away from that socket.

Then, in order to establish the new connection with a target socket,
the user also doesn't need to drag the mouse exactly over the other
socket anymore. As long as the mouse is close enough to the other
socket, it will be detected automatically.

Of course, even if the socket is compatible and close enough, the
connection may still be denied if connecting the sockets is forbidden,
that is, if:

- doing so would cause a cycle in the graph
- the sockets are already connected

Additionally, special graphics/shapes are drawn on the screen when
dragging a new connection to another socket in order to signal the
detection to the user.

The user can even customize the experience by:

- picking one of the following graphics to be used:
  - assisting line
  - reaching hands
  - hands and eyes
  - baseball elements
  - baseball elements and eyes
- setting the distance at which a target socket is detected when the
  mouse is dragging a new connection
- setting the distance at which a connection can be established
  with the detected socket