fix: execution of nodes w/ keyword-variable params
Fixes #100

Nodes with keyword-variable parameters (parameters prefixed with **)
were being executed in the wrong way due to a bug that caused their
arguments to be discarded when more than 01 variable keyword argument
was collected.

This nasty bug could either cause all kinds of errors when executing
the graph or even worse remain undetected, all due to the unexpected
outputs that such nodes produce by ignoring the variable keyword
arguments.

It was introduced in v1.5.0, when the entire graph execution mechanism
was overhauled in order to adopt the topological sorting algorithm.