fix: bug in file manager when cancelling
That is, whenever the user cancelled a path selection session,
if the "expecting files only" flag was on, rather than returning
no paths, the file manager was returning path/paths built with
contents from the entry (if it wasn't empty).

A clearly undesired behaviour, since the user wants to cancel.